requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
chromadb
PyPDF2
python-dotenv
faiss-cpu
google-generativeai
ddgs
vapi_python